Tweaked and fixed various things.
Fixed colors so they work right in black on white terminals.

Added --dry-run tests and output reminding users that --dry-run is being used, for rsync.

Updated history  output to only show history for jobs if last bu was Job, otherwise it says, not a job last bu.

This is all just readability and end user stuff, but makes it a bit nicer overall.